For now, make tokio unconditional and make it include macros feature.
This is due to the issue that the gpu_accel tests require the macros feature to be enabled and I am too lazy to set that conditional up right now. Also, the conditional async impls need to be re-evaluated, epsecially considering gpu fns need to have async synchronization for awaiting gpu task completion via callback fn. This is all tbd.