matmul32_extern_same_size_small now works! See comments for explanation.
I believe the issue lay with the fact that the shader was asking for a vec4<u32> but the buffer being bound (`dims_buffer`) to it had been created init with a slice cast from effectively a [usize; 4] because the individual items put in it were usize by default and were not explicitly cast to u32 before being placed into the transmuted array slice. This would mean that the shader saw the dims buffer as `[lhs.dim().0, 0, lhs.dim().1, 0]` which means that lhs for example would be 3x0 which is why the shader code didn't just misalign but completely bricked bc both arrays always had a y len of 0, meaning 0 total capacity for both.